visualizer/requirements.txt
streamlit==1.8.1
opencv-python
pyyaml
plotly
pandas
